chore: refactor email-verifier tests
